libX11 fontconfig xserver mesa plink xkeyboard-config git update Mar 27 2021
xserver          commit a3931ec6f43857aeed7feac5d223d7db6728145e
xkeyboard-config commit 8d57dc13517e3b6a94bb17bbd4d85075c7ff10da
libX11           commit 5faa8dc0b364f55f19034c370436074646d9ad54
fontconfig       commit b03140c14ed71a8b29ae6d5f76f254a2adfc4f22
mesa             commit b30792730341c3bd20b3dd548a3b47c664834a1e
putty            commit 049acf9ef570f4ed98232d801437c9c1196bfe58

+++ mesalib/src/gallium/frontends/nine/nine_debug.c
@@ -74,7 +74,7 @@ _nine_debug_printf( unsigned long flag,
 
 #if defined(HAVE_PTHREAD)
     if (dbg_flags & DBG_TID)
-        tid = pthread_self();
+        tid = (unsigned long)pthread_self();
 #endif
 
     if (dbg_flags & flag) {
